move SimpleScanStateful
